simplify user api
